first working Python re-implementation of discrete static Gauss model
This is the base for a package of reaction time models that may be fit to data
using ABC-methods. I've started with our discrete static Gauss model, because I
want to investigate how wrong assumptions about nondecision time impact model
inferences.

The major advantage of this implementation over the Matlab version is the large
speedup. I achieved this by using numba for just in time compilation of crucial
parts of the model. The example 'fit_with_EPABC' demonstrates that: Fitting the
model to 200 trials using typical EP-ABC parameters takes ~50s on two of my
computers whereas it takes over 30min when using the Matlab implementation.
Notably, the Matlab implementation ran on 6 cores whereas the Python
implementation is not parallel, yet.

'fit_with_EPABC' also tests differences in inference based on choice and RT vs.
only choice. With the change to ddof=1.5 in EP-ABC I now see that nondecision
time posteriors stay close to the priors when only choice is used, as you should
expect. The major difference in inferred parameters is that the choice-only
version infers lapse probability of ~0.2 whereas the rt-model suggests lapse
probability close to 0. At the same time, the choice-only model inferred a lower
noisestd. The question is whether the rt-model is deceived by wrong assumptions
about RT?