new model (sdsg) + bugfix
new model: sensory discrete static Gauss model (sdsg)
This inherets from discrete static Gauss model and only adds one more parameter,
sensdrift. The new model implements the idea that each individual, sequentially
presented feature needs to be recognised before it can contribute to the
decision. I realise this by using a simple drift diffusion process to determine
the contribution of each feature to the overall decision. The state of the
drift diffusion process after some time dt is simply a Gaussian with a mean
depending on the drift and a variance depending on the diffusion. So I simply
sample from the Gaussian to determine the contribution of that feature to the
overall decision. Importantly, I limit the mean of the Gaussian to the actual
log-likelihood ratio for this feature. This is the only place where the feature
enters the computations. Two consequences: 1) For very high sensdrift the model
will always reach the limit, i.e., log-likelihood ratio, implementing the
standard discrete static Gauss model. 2) For very low sensdrift the model will
never reach the limit and will implement a standard drift diffusion model. As a
result, the model will only be useful in the intermediate regime where sometimes
the limit is reached and sometimes not, i.e., the value of dt*sensdrift must
somehow be in the range of the expected log-likelihood ratios.

implementation note:
To do as little as possible code copying I introduced a new function
gen_response_jitted which calls the model specific jitted functions. This way it
was possible to let the new model inherit from discrete_static_gauss and only
implement gen_response_jitted and __init__.

bugfix:
Trials.setter was not accessing shape of Trials when using features