fixing major bugs; dsg now consistent with Matlab implementation; tile
major bug 1:
Models predicted too short RTs (by one dt), because I didn't take into account
that counter 't' starts from 0, not 1.

major bug 2:
I accidentally scrambled the order of predicted responses in
gen_response_from_Gauss_posterior.

After fixing these bugs the results of fitting the discrete static Gauss model
(dsg) are consistent with those of the Matlab implementation, i.e., the
resulting posteriors, log marginal likelihoods and posterior predictive log
likelihoods were quite close (deviations could be explained by variation across
runs of the fitting) for the tested subject (11) and condition (25) of the Bee
experiment, see also test_sensory_gauss_model.py in BeeExpAnalysis. This holds
for the model when it uses the dots. Another check for when it is not using the
dots is currently running, but I don't expect any upsets.

There still seems to be something wrong with the sensory discrete static Gauss
model: When using extremely large sensdrift, the resulting posteriors should be
similar to the dsg, because for every shown dot the corresponding log likelihood
ratio should be reached, but this is not what happened when I moved the prior
for sensdrift to large values.

small change in implementation:
use tile instead of kron to repeat arrays