major bug fix in sdsg: model works, is faster than dsg
The sensory discrete static gauss model did not work, because I computed the log
likelihood ratio wrongly: I used noisestd instead of the internal uncertainty
(variance).

With the fix the model works nicely. For one data set (see BeeExpAnalysis) I
have verified that sdsg recovers dsg if the prior for sensdrift only allows very
large values such that the llr is always reached. I could not confirm my other
intuition: that for very small sensdrift sdsg implements a standard DDM, because
then the dot llr becomes irrelevant, because it is never reached. Neither the
variant where at least the sign of sensdrift follows the sign of the dot llr,
nor the variant where the dot positions are completely unused (sign of sensdrift
determined by mean dot position for that trial as in the DDM) led to similarly
good predictions as using the dsg. The reason may have been that the inferred
posterior was wide and had relatively large nondecision times. Perhaps this is
exactly the problem that you will see when the model is ill-defined (as for the
DDM, or sdsg with small sensdrift where drift, diffusion and bound are all
allowed to vary).

Inference with sdsg is approximately twice as fast as inference with dsg. This
is probably because sdsg runs directly on log likelihood ratios (2-choice)
whereas dsg implements also multiple choice decisions and, hence, renormalises
posteriors after each time step.