bugfix: make prior_re more strict
The regular expression which identifies the prior in a parameter name was a bit
too flexible: It also matched, e.g., "priorMean" which Pouyan used in his eBM
such that "priorMean" was internally converted into "prior".

The "$" at the end of the re now restricts matches to strings which either are
"prior" or "prior_d" where d is a digit. This change should minimise unwanted
interactions with other parameter names.