bugfix and robustness for prior
These changes were triggered by a problem that Pouyan ran into when simulating
from the model while choosing a particular, single value for the prior by giving
that value in a pardict to gen_response_with_params. So far I did not check
whether the prior is in the right format (numpy array with C-1 elements) in that
case. I now have included a check which automatically transforms a scalar to an
array, as is done when parameters are not given as dictionary.

Additionally, I now have introduced appropriate checks for the prior by means of
setter and getter methods in the model.

extra bugfix: When only 1 sample from the model was requested (1 given trial and
1 given set of parameters) an error occurred, because the scalar parameter
values were not automatically transformed into arrays. This is now done.