introduced coupling switch + new formula for coupling
Having realised that you probably need coupling between noisestd and intstd to
get slow errors I here introduce a switch to use or not to use coupling in the
model as a model parameter.

There were some scaling issues when computing the coupled intstd. This is
because the coupled intstd depends on the setting of the means and on the
diffusion in the drift diffusion model.

diffusion: although the diffusion s is set to 0.1 in the DDM, we effectively
reset it by assuming posterior bound values around 0.8. Given that typical DDM
bounds lead to posterior bounds of 0.51 I computed the scaling factor to get the
bound from 0.51 to 0.8 as 34.66. The diffusion s then must be scaled by the
square root of that factor giving a value around 3.5 which I now use in the
computation of the coupled intstd.

means: Previously we used a factor of 2 in the coupling equation. The 2 derived
from the assumption that the means are -1 and 1. Properly deriving the equations
for multidimensional means shows that the factor should be the Euclidean
distance between the two means which I now implemented.