switch off cache for gen_response_jitted_dsg because of numba bug
Testing on my laptop I found that the first call to gen_response_jitted_dsg in a
new ipython session will break with a KeyError produced by numba's caching
mechanism. This only happens for gen_response_jitted_dsg, not for other jitted
functions such as normaliselogprob. It seems that this numba issue is related:

https://github.com/numba/numba/issues/1466

The error only affects the cache and can be prevented by not caching
gen_response_jitted_dsg. Apparently the bug is also fixed in numba 0.29. I
currently don't know why I have not come across this bug on computers in the
office.