new implementation of discrete static Gauss model for computing intermediates
I implemented new methods which return different quantities that describe the
evidence accumulation process. Specifically I compute log likelihoods and log
posteriors in compute_logpost_from_features() and I compute surprise in
compute_surprise(). The discrete static Gauss model originally uses the log
posteriors to make decisions and this can be emulated by passing the returned
log posteriors into gen_response_from_logpost().

why new pure python implementation?:
Because I expect these methods to be used in contexts where speed is not a
priority, I implemented them in pure python. I originally hesitated to do that,
because I duplicate code this way. The alternative would have been to separate
all the components of gen_response_jitted_dsg() into separate functions, but I
feared that this would compromise the speed of this implementation. Also, the
logic of gen_response_jitted is different: it immediately stops simulating when
a bound is hit, but the focus of the new code is to compute the intermediate log
posteriors and likelihoods.

consistency:
I have also written examples/test_consistency_dsg.py which collects basic tests
which can be used to test for deviations between the numba and new pure python
implementations of the model. The main rationale is that responses generated
from both implementations should be equal, if there is no randomness in the
generation of responses. For the parameter values I tested I could confirm that
this is indeed the case, i.e., that the two implementations are consistent.

changing normaliselogprob to logsumexp:
The new implementation showed that it would be better (more general) to just
compute the log of the normalisation constant and not immediately subtract it
from the log probabilities. So I changed normaliselogprob to logsumexp which is
the same as numpy's logaddexp, but for more than two values. I also wanted to
make a version which takes in arrays with arbitrary dimension and computes the
log normalising constants along a given dimension, but this cannot be done in
numba, because it doesn't support the 'axis' arguments of the numpy functions I
use. It's also not easy/possible to implement such code directly in numba which
is why I decided to at least make a 3D version in which the axis over which to
compute the log normalising constant can be chosen.

computation of surprise:
I followed my original Matlab implementation, but found a bug there which I
corrected here: The log posterior at time t was combined with the likelihoods of
the same time point, but it should be the log posterior at time t-1 which should
be added to the log likelihoods at time t. I checked that the resulting values
of surprise are correct by comparing them to the likelihoods of the multivariate
normal for the case that the log-posteriors put all mass on one choice, see
examples/test_consistency_dsg.py