major bugfix: class vs instance objects + some new defaults + cleanup
I did not properly acknowledge the difference between class and instance objects
in my old implementation and defined model parameters as class objects. This is
dangerous, because class objects will have the same value for all instances of
the class and changing the value of a class object from one instance will change
the values of all other instances.

Here I have made all parameters and other instance-specific variables such as
the number of feature dimensions or the number of trials stored in the model
into instance objects.

During the transformation I have made other changes that came to my attention:
- new default values:
	* ndtmean of dsg: -12 to make ndt effectively 0
	* dt: 1
	* toresponse: [0, maxrt+1]
- dsg means will now be transformed to np.array automatically
- pythonised function arguments: proper use of default values and **kwargs
- use_features in dsg is now read-only and is determined from the dimensionality
  of Trials