bugfix: computed surprise of first dot had unwanted offset
I overlooked that the assignment

log_post = log_liks

makes log_post only a view of the array pointed to by log_liks. By then changing
the first elements of log_post to add the contribution of the prior I also
updated log_liks which was unwanted. This had no effect on the correct
computation of log_post, but it meant that subsequent surprise for the first dot
contained the contribution of the prior twice (once from log_liks and once from
log_post which becomes its own array after the cumsum). I now make an explicit
copy of log_liks which prevents this behaviour.